+++ android/app/src/main/java/com/targetinduction/MainActivity.java
@@ -1,4 +1,4 @@
-package com.reactnativebase;
+package com.targetinduction;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -9,7 +9,7 @@ public class MainActivity extends ReactActivity {
 
   @Override
   protected String getMainComponentName() {
-    return "ReactNativeBase";
+    return "targetInduction";
   }
 
   @Override
