+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'ReactNativeBase'
+rootProject.name = 'targetInduction'
 include ':@react-native-community_async-storage'
 project(':@react-native-community_async-storage').projectDir = new File(rootProject.projectDir, '../node_modules/@react-native-community/async-storage/android')
