+++ ios/Podfile
@@ -34,12 +34,12 @@ def base_pods
   pod 'Folly', :podspec => '../node_modules/react-native/third-party-podspecs/Folly.podspec'
 end
 
-target 'ReactNativeBase' do
-  # Pods for ReactNativeBase
+target 'targetInduction' do
+  # Pods for targetInduction
   base_pods
 
 
-  target 'ReactNativeBaseTests' do
+  target 'targetInductionTests' do
     inherit! :search_paths
     # Pods for testing
   end
@@ -47,26 +47,26 @@ target 'ReactNativeBase' do
   use_native_modules!
 end
 
-target 'ReactNativeBase-Staging' do
-    # Pods for ReactNativeBase
+target 'targetInduction-Staging' do
+    # Pods for targetInduction
     base_pods
 
 
     use_native_modules!
 end
 
-target 'ReactNativeBase-Develop' do
-  # Pods for ReactNativeBase
+target 'targetInduction-Develop' do
+  # Pods for targetInduction
   base_pods
 
 
   use_native_modules!
 end
 
-target 'ReactNativeBase-tvOS' do
-  # Pods for ReactNativeBase-tvOS
+target 'targetInduction-tvOS' do
+  # Pods for targetInduction-tvOS
 
-  target 'ReactNativeBase-tvOSTests' do
+  target 'targetInduction-tvOSTests' do
     inherit! :search_paths
     # Pods for testing
   end
