+++ ios/targetInduction.xcodeproj/xcshareddata/xcschemes/targetInduction.xcscheme
@@ -14,9 +14,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D2A28121D9B038B00D4039D"
+               BlueprintIdentifier = "83CBBA2D1A601D0E00E9B192"
                BuildableName = "libReact.a"
-               BlueprintName = "React-tvOS"
+               BlueprintName = "React"
                ReferencedContainer = "container:../node_modules/react-native/React/React.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -28,10 +28,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBase-tvOS.app"
-               BlueprintName = "ReactNativeBase-tvOS"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+               BuildableName = "targetInduction.app"
+               BlueprintName = "targetInduction"
+               ReferencedContainer = "container:targetInduction.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -42,10 +42,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBase-tvOSTests.xctest"
-               BlueprintName = "ReactNativeBase-tvOSTests"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
+               BuildableName = "targetInductionTests.xctest"
+               BlueprintName = "targetInductionTests"
+               ReferencedContainer = "container:targetInduction.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -55,29 +55,27 @@
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
+      <MacroExpansion>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "targetInduction.app"
+            BlueprintName = "targetInduction"
+            ReferencedContainer = "container:targetInduction.xcodeproj">
+         </BuildableReference>
+      </MacroExpansion>
       <Testables>
          <TestableReference
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "ReactNativeBase-tvOSTests.xctest"
-               BlueprintName = "ReactNativeBase-tvOSTests"
-               ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+               BlueprintIdentifier = "00E356ED1AD99517003FC87E"
+               BuildableName = "targetInductionTests.xctest"
+               BlueprintName = "targetInductionTests"
+               ReferencedContainer = "container:targetInduction.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBase-tvOS.app"
-            BlueprintName = "ReactNativeBase-tvOS"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
-      <AdditionalOptions>
-      </AdditionalOptions>
    </TestAction>
    <LaunchAction
       buildConfiguration = "Debug"
@@ -93,14 +91,12 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBase-tvOS.app"
-            BlueprintName = "ReactNativeBase-tvOS"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "targetInduction.app"
+            BlueprintName = "targetInduction"
+            ReferencedContainer = "container:targetInduction.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
-      <AdditionalOptions>
-      </AdditionalOptions>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
@@ -112,10 +108,10 @@
          runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "ReactNativeBase-tvOS.app"
-            BlueprintName = "ReactNativeBase-tvOS"
-            ReferencedContainer = "container:ReactNativeBase.xcodeproj">
+            BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
+            BuildableName = "targetInduction.app"
+            BlueprintName = "targetInduction"
+            ReferencedContainer = "container:targetInduction.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
