+++ ios/targetInduction/Images.xcassets/AppIcon.appiconset/Contents.json
